datasets/councils/pyproject.toml
# package build
# https://peps.python.org/pep-0517/
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

# package basics
# https://peps.python.org/pep-0621/
[project]
name = "lab-scale-ai-data-councils"
description = "Dataset manager for the Councils in Action dataset"
keywords = []
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT License" }
authors = [
  { email = "evamxb@gmail.com", name = "Eva Maxfield Brown" },
]
version = "0.1.0"
classifiers = [
  "Development Status :: 4 - Beta",
  "Natural Language :: English",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
]
dependencies = [
  "cdp-data==0.0.10",
  "datasets>=2,<3",
  "pandas",  # comes from cdp-data
  "pyarrow>=9.0.0",
  "python-dotenv>=1,<2",
  "scikit-learn>=1,<2",
  "tqdm",
  "typer>=0.9.0,<1",
]

[project.urls]
Homepage = "https://github.com/BillHoweLab/lab-scale-ai"
"Bug Tracker" = "https://github.com/BillHoweLab/lab-scale-ai/issues"
Documentation = "https://github.com/BillHoweLab/lab-scale-ai"
"User Support" = "https://github.com/BillHoweLab/lab-scale-ai/issues"

# extra dependencies
# https://peps.python.org/pep-0621/#dependencies-optional-dependencies
[project.optional-dependencies]
lint = [
  "pre-commit>=2.20.0",
]
dev = [
  "ipython",
  "jupyterlab",
]

# entry points
# https://peps.python.org/pep-0621/#entry-points
[project.entry-points."console_scripts"]
manage-councils-dataset = "src.main:main"

# build settings
# https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]
include = ["src"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

# tools
[tool.black]
line-length = 88
target-version = ["py311"]

# https://github.com/charliermarsh/ruff
[tool.ruff]
line-length = 88
target-version = "py311"
extend-select = [
    "E",      # style errors
    "F",      # flakes
    "D",      # pydocstyle
    "I001",   # isort
    "UP",     # pyupgrade
    "N",      # pep8-naming
    # "S",    # bandit
    "C",      # flake8-comprehensions
    "B",      # flake8-bugbear
    "A001",   # flake8-builtins
    "RUF",    # ruff-specific rules
    "RUF100", # Unused noqa directive
]
extend-ignore = [
    "D100", # Missing docstring in public module
    "D101", # Missing docstring in public class
    "D103", # Missing docstring in public function
    "D107", # Missing docstring in __init__
    "D203", # 1 blank line required before class docstring
    "D205", # 1 blank line required between summary line and description
    "D212", # Multi-line docstring summary should start at the first line
    "D213", # Multi-line docstring summary should start at the second line
    "D413", # Missing blank line after last section
    "D416", # Section name should end with a colon
]

[tool.ruff.per-file-ignores]
"**/notebooks/*" = ["E501"]
"**/tests/*.py" = ["D"]

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true
show_error_codes = true